+++ src/memex_logging/ws/resource/analytic.py
@@ -23,7 +23,10 @@ from flask import request, Response
 from flask_restful import Resource, abort
 
 from memex_logging.common.analytic.builder import AnalyticBuilder
-from memex_logging.celery.analytic import compute_analytic
+from memex_logging.celery.analytic import update_analytic
+from memex_logging.common.model.aggregation import AggregationAnalytic
+from memex_logging.common.model.analytic import DimensionAnalytic
+from memex_logging.common.model.response import AnalyticResponse, AggregationResponse
 from memex_logging.common.utils import Utils
 
 
@@ -49,7 +52,7 @@ class AnalyticInterface(Resource):
         static_id = request.args.get('staticId')
         logger.info(f"Retrieving analytic with static_id: {static_id}")
         if static_id == "" or static_id is None:
-            logger.info("Missing required staticId parameter")
+            logger.debug("Missing required staticId parameter")
             return {
                 "status": "Malformed request: missing required parameter `staticId`",
                 "code": 400
@@ -79,7 +82,7 @@ class AnalyticInterface(Resource):
         analytic = request.json
         logger.info(f"Creating analytic: {analytic}")
         if analytic is None:
-            logger.info("Analytic failed to be computed due to missing data")
+            logger.debug("Analytic failed to be computed due to missing data")
             return {
                 "status": "Malformed request: data is missing",
                 "code": 400
@@ -88,14 +91,37 @@ class AnalyticInterface(Resource):
         try:
             analytic = AnalyticBuilder.from_repr(analytic)
         except (KeyError, ValueError, TypeError, AttributeError) as e:
-            logger.debug("Error while parsing input analytic data", exc_info=e)
+            logger.warning("Error while parsing input analytic data", exc_info=e)
             return {
                 "status": f"Malformed request: analytic not valid. Cause: {e.args[0]}",
                 "code": 400
             }, 400
+        except Exception as e:
+            logger.exception("Something went wrong in parsing the analytic", exc_info=e)
+            return {
+                "status": "Internal server error: something went wrong in parsing the analytic",
+                "code": 500
+            }, 500
 
         static_id = str(uuid.uuid4())
-        compute_analytic.delay(raw_analytic=analytic.to_repr(), static_id=static_id)
+        if isinstance(analytic, DimensionAnalytic):
+            index_name = "analytic-" + analytic.project.lower() + "-" + analytic.dimension.lower()
+            self._es.index(index=index_name, doc_type='_doc', body=AnalyticResponse(analytic, None, static_id).to_repr())
+            logger.debug(f"Analytic stored in index [{index_name}]")
+
+        elif isinstance(analytic, AggregationAnalytic):
+            index_name = "analytic-" + analytic.project.lower() + "-" + analytic.aggregation.lower()
+            self._es.index(index=index_name, doc_type='_doc', body=AggregationResponse(analytic, None, static_id).to_repr())
+            logger.debug(f"Aggregation stored in index [{index_name}]")
+
+        else:
+            logger.error(f"Unrecognized class of analytic [{type(analytic)}]")
+            return {
+                "status": "Internal server error: something went wrong in handling the analytic",
+                "code": 500
+            }, 500
+
+        update_analytic.delay(static_id=static_id)
         return {"staticId": static_id}, 200
 
     def delete(self):