+++ test/unit/memex_logging/common/dao/test_message.py
@@ -17,13 +17,15 @@ from __future__ import absolute_import, annotations
 from datetime import datetime
 from unittest import TestCase
 
+from elasticsearch import Elasticsearch
+
 from memex_logging.common.dao.message import MessageDao
 
 
 class TestMessageDao(TestCase):
 
     def test_build_query_based_on_parameters(self):
-        message_dao = MessageDao(None)
+        message_dao = MessageDao(Elasticsearch())
 
         query = message_dao._build_query_based_on_parameters(trace_id="trace_id",  message_id=None, user_id=None)
         self.assertEqual({
@@ -72,13 +74,10 @@ class TestMessageDao(TestCase):
             message_dao._build_query_based_on_parameters(trace_id=None, message_id=None, user_id="user_id")
 
     def test_generate_index_based_on_time_range(self):
-        message_dao = MessageDao(None)
+        message_dao = MessageDao(Elasticsearch())
 
         with self.assertRaises(ValueError):
             message_dao._generate_index_based_on_time_range("project", datetime(2021, 2, 6), datetime(2021, 2, 5))
-
-        index = message_dao._generate_index_based_on_time_range("project", datetime(2021, 2, 5), datetime(2021, 2, 5))
-        self.assertEqual(f"{MessageDao.BASE_INDEX}-project-2021-02-05", index)
 
         index = message_dao._generate_index_based_on_time_range("project", datetime(2021, 2, 4), datetime(2021, 2, 5))
         self.assertEqual(f"{MessageDao.BASE_INDEX}-project-*", index)